Fixed the login error.
The login_successful variable I used before was apparently only being
passed in when it was successful (even though it was being set to false
on failure.  So this caused the conditional `if not login_successful`
to always be true.  I fixed this by passing in a login_failed value
only when it fails.