Committing terrible code.
I made the course_grade page have a form input for each student
enrolled in the course.  The input has the initial value of the
student's grade and when a grade is changed and a post is made the
enrolled_in is updated.  Still needs min/max limit and pretty-ing up.