Make minor changes
Changed the bucket list buttons to say cart instead.
Added start and end times to course info.
Changed "Days & Times" label to "Sessions" (@njc3006).  Sessions was
shorter which looked nicer on the course info page.  But if you
decide to change it back use the escaped form of the ampersand.